% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{api_get_request}
\alias{api_get_request}
\title{Wrapper of httr::GET that handles status errors and custom headers}
\usage{
api_get_request(endpoint, params, call = sys.call(-1))
}
\arguments{
\item{endpoint}{An EpiGraphDB API endpoint, e.g. "/mr"}

\item{params}{GET request params}

\item{call}{The function call to identify the original function
when things go wrong}
}
\description{
Wrapper of httr::GET that handles status errors and custom headers
}
\keyword{internal}
