% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{api_request}
\alias{api_request}
\title{The very general wrapper from EpiGraphDB endpoint request}
\usage{
api_request(
  route,
  params,
  mode = c("table", "raw"),
  method = api_get_request,
  retry_times,
  retry_pause_min,
  call = sys.call(-1)
)
}
\arguments{
\item{route}{An EpiGraphDB API endpoint route, e.g. "/mr"}

\item{params}{A list of parameters to send}

\item{mode}{Either \code{"table"} (returns tibble) or
\code{"raw"} (returns raw response parsed from json to R list).}

\item{method}{A specific request handler, e.g. \code{epi_get_request}}

\item{call}{The function call to identify the original function
when things go wrong}
}
\description{
The very general wrapper from EpiGraphDB endpoint request
}
\keyword{internal}
