% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypher.R
\name{cypher}
\alias{cypher}
\title{Send a query in Cypher to EpiGraphDB}
\usage{
cypher(query, mode = c("table", "raw"))
}
\arguments{
\item{query}{A Cypher query.}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{\code{tibble}} as per
\href{https://style.tidyverse.org/}{\code{tidyverse}} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{\code{httr}}.}
}
\description{
NOTE: this function is intended for advanced uses.
Regular users are recommended to use standard query functions
}
\examples{
cypher("MATCH (n:Gwas) RETURN n LIMIT 2")
}
