% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{flatten_response}
\alias{flatten_response}
\title{Flatten the "results" field from an API response to a tibble df}
\usage{
flatten_response(response, field = "results")
}
\arguments{
\item{response}{An httr response}

\item{field}{Default to the "results" field}
}
\value{
A tibble df
}
\description{
The general tibble flattener for EpiGraphDB endpoints
}
\keyword{internal}
