% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_cor.R
\name{gwas_cor}
\alias{gwas_cor}
\title{GWAS correlations between traits}
\usage{
gwas_cor(trait, cor_coef_threshold = 0.8, mode = c("table", "raw"))
}
\arguments{
\item{trait}{name of the trait, e.g. "body mass index"}

\item{cor_coef_threshold}{correlation coefficient threshold}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{tibble} as per
\href{https://style.tidyverse.org/}{tidyverse} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{httr}.}
}
\value{
Data from \code{gwas_cor}
}
\description{
GWAS correlations between traits
}
\examples{
gwas_cor(trait = "Body mass index") \%>\%
  dplyr::glimpse()

# Use a different threshold
gwas_cor(trait = "Body mass index", cor_coef_threshold = 0.4) \%>\%
  dplyr::glimpse()

}
