% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwas_cor.R
\name{gwas_cor_requests}
\alias{gwas_cor_requests}
\title{Requests /gwas_cor}
\usage{
gwas_cor_requests(trait, cor_coef_threshold)
}
\arguments{
\item{trait}{name of the trait, e.g. "body mass index"}

\item{cor_coef_threshold}{correlation coefficient threshold}
}
\description{
Requests /gwas_cor
}
\keyword{internal}
