% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{http_condition}
\alias{http_condition}
\title{Modified httr::http_condition}
\usage{
http_condition(response, call = sys.call(-1))
}
\arguments{
\item{response}{An httr response}

\item{call}{The function call}
}
\description{
Modified httr::http_condition
}
\keyword{internal}
