% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/literature.R
\name{literature_mr}
\alias{literature_mr}
\title{Literature evidence regarding MR}
\usage{
literature_mr(trait_name, semmed_predicate = NULL, mode = c("table",
  "raw"))
}
\arguments{
\item{trait_name}{A trait name}

\item{semmed_predicate}{Either NULL which returns entries from
all predicates, or a SemMed predicate e.g. "DIAGNOSES" or "ASSOCIATED_WITH"}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{tibble} as per
\href{https://style.tidyverse.org/}{tidyverse} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{httr}.}
}
\value{
Data from
\href{http://api.epigraphdb.org/#/literature/get_literature_mr_literature_mr_get}{ /literature/mr }
}
\description{
Literature evidence regarding MR
}
\examples{
literature_mr(trait_name = "Body mass index")
}
