% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_rels_list}
\alias{meta_rels_list}
\title{List meta rels (e.g. MR, etc.)}
\usage{
meta_rels_list(mode = c("raw"))
}
\arguments{
\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{\code{tibble}} as per
\href{https://style.tidyverse.org/}{\code{tidyverse}} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{\code{httr}}.}
}
\value{
Data from \code{GET /meta/rels/list}
}
\description{
\href{https://docs.epigraphdb.org/api/api-endpoints/#get-metarelslist}{\code{GET /meta/rels/list}}
}
\examples{
meta_rels_list()
}
