% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr}
\alias{mr}
\title{Return information related to Mendelian Randomisation}
\usage{
mr(exposure = NULL, outcome = NULL, pval_threshold = 1e-05,
  mode = c("table", "raw"))
}
\arguments{
\item{exposure}{A trait name, eg. "Body mass index",
leaving \code{exposure} as \code{NULL} will return MR information
related to a specific \code{outcome}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}

\item{outcome}{A trait name, eg. "Coronary heart disease",
leaving \code{outcome} as \code{NULL} will return MR information
related to a specific \code{exposure}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}

\item{pval_threshold}{P-value threshold for the MR evidence.}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{tibble} as per
\href{https://style.tidyverse.org/}{tidyverse} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{httr}.}
}
\value{
Data from
\href{http://devapi.epigraphdb.org/#/topics/get_mr_mr_get}{/mr}
}
\description{
Return information related to Mendelian Randomisation
}
\examples{
# Returns a data frame
mr(exposure = "Body mass index", outcome = "Coronary heart disease")

# Returns raw response
mr(
  exposure = "Body mass index", outcome = "Coronary heart disease",
  mode = "raw"
) \%>\% str()

# Use a different threshold
mr(exposure = "Body mass index", pval_threshold = 1e-8)
}
