% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_regulator}
\alias{mr_regulator}
\title{Regulate parameter input}
\usage{
mr_regulator(exposure, outcome)
}
\arguments{
\item{exposure}{A trait name, eg. "Body mass index",
leaving \code{exposure} as \code{NULL} will return MR information
related to a specific \code{outcome}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}

\item{outcome}{A trait name, eg. "Coronary heart disease",
leaving \code{outcome} as \code{NULL} will return MR information
related to a specific \code{exposure}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}
}
\description{
Regulate parameter input
}
\keyword{internal}
