% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_requests}
\alias{mr_requests}
\title{The core function to query from epigraphdb api}
\usage{
mr_requests(exposure, outcome, pval_threshold)
}
\arguments{
\item{exposure}{A trait name, eg. "Body mass index",
leaving \code{exposure} as \code{NULL} will return MR information
related to a specific \code{outcome}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}

\item{outcome}{A trait name, eg. "Coronary heart disease",
leaving \code{outcome} as \code{NULL} will return MR information
related to a specific \code{exposure}.
\strong{NOTE}: \code{exposure} and \code{outcome} cannot be both \code{NULL}.}

\item{pval_threshold}{P-value threshold for the MR evidence.}
}
\description{
The core function to query from epigraphdb api
}
\keyword{internal}
