% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_table}
\alias{mr_table}
\title{Reformat reponse from mr into a table}
\usage{
mr_table(response)
}
\arguments{
\item{response}{response from mr}
}
\description{
Reformat reponse from mr into a table
}
\keyword{internal}
