% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_cor.R
\name{obs_cor}
\alias{obs_cor}
\title{Observational correlations between traits}
\usage{
obs_cor(trait, pval_threshold = 1e-05, mode = c("table", "raw"))
}
\arguments{
\item{trait}{name of the trait, e.g. "body mass index"}

\item{pval_threshold}{pvalue threshold}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{tibble} as per
\href{https://style.tidyverse.org/}{tidyverse} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{httr}.}
}
\value{
Data from
\href{http://api.epigraphdb.org/#/topics/get_obs_cor_obs_cor_get}{ /obs_cor }
}
\description{
Observational correlations between traits
}
\examples{
obs_cor(trait = "Body mass index (BMI)") \%>\%
  dplyr::glimpse()

# Use a different threshold
obs_cor(trait = "Body mass index (BMI)", pval_threshold = 1e-8) \%>\%
  dplyr::glimpse()
}
