% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_cor.R
\name{obs_cor_requests}
\alias{obs_cor_requests}
\title{Requests /obs_cor}
\usage{
obs_cor_requests(trait, pval_threshold)
}
\arguments{
\item{trait}{name of the trait, e.g. "body mass index"}

\item{pval_threshold}{pvalue threshold}
}
\description{
Requests /obs_cor
}
\keyword{internal}
