% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{ontology}
\alias{ontology}
\title{Ontology association between EFO term and Gwas}
\usage{
ontology(efo_term, mode = c("table", "raw"))
}
\arguments{
\item{efo_term}{EFO term, e.g. "systolic blood pressure"}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{tibble} as per
\href{https://style.tidyverse.org/}{tidyverse} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{httr}.}
}
\value{
Data from
\href{http://api.epigraphdb.org/#/topics/get_ontology_ontology_get}{ ontology }
}
\description{
Ontology association between EFO term and Gwas
}
\examples{
ontology(efo_term = "systolic blood pressure")
}
