% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{ontology_requests}
\alias{ontology_requests}
\title{Requests /ontology}
\usage{
ontology_requests(efo_term)
}
\arguments{
\item{efo_term}{EFO term, e.g. "systolic blood pressure"}
}
\description{
Requests /ontology
}
\keyword{internal}
