% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{ontology_table}
\alias{ontology_table}
\title{Table format /ontology}
\usage{
ontology_table(response)
}
\arguments{
\item{response}{response from \code{ontology_requests}}
}
\description{
Table format /ontology
}
\keyword{internal}
