% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_list_requests}
\alias{pqtl_list_requests}
\title{Create an API request for the list of proteins/traits}
\usage{
pqtl_list_requests(flag)
}
\arguments{
\item{flag}{(Optional) A flag which indicates whether the list of
exposures (if "exposures") or outcomes (if "outcomes")
should be returned. The DEFAULT is "exposures".}
}
\description{
Create an API request for the list of proteins/traits
}
\keyword{internal}
