% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_pleio_regulator}
\alias{pqtl_pleio_regulator}
\title{Regulate parameter input}
\usage{
pqtl_pleio_regulator(rsid, prflag)
}
\arguments{
\item{rsid}{(Required) A SNP identified by rsID which cannot be \code{NULL}.}

\item{prflag}{(Optional) A flag which determines whether the number (if "count")
or names (if "proteins") of the associated proteins should be returned.
The DEFAULT value is "proteins".}
}
\description{
Regulate parameter input
}
\keyword{internal}
