% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_pleio_requests}
\alias{pqtl_pleio_requests}
\title{Create an API request for the pleiotropy analyses}
\usage{
pqtl_pleio_requests(rsid, prflag)
}
\arguments{
\item{prflag}{}
}
\description{
Create an API request for the pleiotropy analyses
}
\keyword{internal}
