% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_regulator}
\alias{pqtl_regulator}
\title{Regulate parameter input}
\usage{
pqtl_regulator(query, rtype, pvalue, searchflag)
}
\arguments{
\item{searchflag}{}
}
\description{
Regulate parameter input
}
\keyword{internal}
