% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_requests}
\alias{pqtl_requests}
\title{Create an API request for the pQTL-related analyses}
\usage{
pqtl_requests(query, rtype, pvalue, searchflag)
}
\arguments{
\item{searchflag}{}
}
\description{
Create an API request for the pQTL-related analyses
}
\keyword{internal}
