% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_table}
\alias{pqtl_table}
\title{Reformat response from pQTL results into a table}
\usage{
pqtl_table(response)
}
\arguments{
\item{response}{response for pQTL analysis}
}
\description{
Reformat response from pQTL results into a table
}
\keyword{internal}
