% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{stop_for_status}
\alias{stop_for_status}
\title{Catch error for an HTTP response}
\usage{
stop_for_status(response, context)
}
\arguments{
\item{response}{An httr response}

\item{context}{A list on the url and params for the request}
}
\description{
Modifies from httr::stop_for_status
}
\keyword{internal}
