% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{api_post_request}
\alias{api_post_request}
\title{Wrapper of httr::POST that handles status errors and custom headers}
\usage{
api_post_request(route, params, retry_times, retry_pause_min)
}
\arguments{
\item{route}{An EpiGraphDB API endpoint route, e.g. "/mr"}

\item{params}{POST request payload}
}
\description{
Wrapper of httr::POST that handles status errors and custom headers
}
\keyword{internal}
