% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraphdb-package.R
\docType{package}
\name{epigraphdb-package}
\alias{epigraphdb}
\alias{epigraphdb-package}
\title{epigraphdb: EpiGraphDB}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R package for getting EpiGraphDB data.
}
\author{
\strong{Maintainer}: Yi Liu \email{yi6240.liu@bristol.ac.uk}

Authors:
\itemize{
  \item Valeriia Haberland \email{valeriia.haberland@bristol.ac.uk}
  \item Pau Erola \email{pau.erola@bristol.ac.uk}
  \item Benjamin Elsworth \email{ben.elsworth@bristol.ac.uk}
  \item Tom Gaunt \email{Tom.Gaunt@bristol.ac.uk}
}

}
\keyword{internal}
