% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigraphdb-package.R
\docType{package}
\name{epigraphdb-package}
\alias{epigraphdb}
\alias{epigraphdb-package}
\title{epigraphdb: EpiGraphDB}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

What the package does (one paragraph).
}
\author{
\strong{Maintainer}:  EpiGraphDB team \email{EpiGraphDB@epigraphdb.org}

}
\keyword{internal}
