% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_and_response.R
\name{http_condition}
\alias{http_condition}
\title{Modified httr::http_condition}
\usage{
http_condition(response, context)
}
\arguments{
\item{response}{An httr response}

\item{context}{A list on the url and params for the request}
}
\description{
Modified httr::http_condition
}
\keyword{internal}
