% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.R
\name{mr_requests}
\alias{mr_requests}
\title{The core function to query from epigraphdb api}
\usage{
mr_requests(exposure, outcome, pval_threshold)
}
\arguments{
\item{outcome}{}
}
\description{
The core function to query from epigraphdb api
}
\keyword{internal}
