% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.R
\name{pathway}
\alias{pathway}
\title{Pathway evidence}
\usage{
pathway(trait, pval_threshold = 1e-05, mode = c("table", "raw"))
}
\arguments{
\item{trait}{A trait name}

\item{pval_threshold}{P-value threshold}

\item{mode}{If \code{mode = "table"}, returns a data frame
(a \href{https://tibble.tidyverse.org/}{\code{tibble}} as per
\href{https://style.tidyverse.org/}{\code{tidyverse}} convention).
If \code{mode = "raw"}, returns a raw response from EpiGraphDB API
with minimal parsing done by \href{https://httr.r-lib.org/}{\code{httr}}.}
}
\value{
Data from \code{GET /pathway}
}
\description{
\href{https://docs.epigraphdb.org/api/api-endpoints/#get-pathway}{\code{GET /pathway}}
}
\examples{
\dontrun{
pathway(trait = "Body mass index")
}
}
