% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pqtl.R
\name{pqtl_pleio_regulator}
\alias{pqtl_pleio_regulator}
\title{Regulate parameter input}
\usage{
pqtl_pleio_regulator(rsid, prflag)
}
\arguments{
\item{prflag}{}
}
\description{
Regulate parameter input
}
\keyword{internal}
